Rest api has been adapted in order to use pods
The rest api can now read and write in pods.
Database has been changed to mongodb, since it is easier to manage in typescript, and it is posible to have the database running in mongo db servers.
TODO: In order to find reviews for a given place the database must be queries to know where the reviews are.
TODO: Read/Write permissions to the folders of the pod must be properly configured.